Renaming constants
